Add support for arm64
